Rename school uc files
