+++ apps/server/src/modules/school/uc/index.ts
@@ -1,0 +1,1 @@
+export * from './legacy-school.uc';
