+++ apps/server/src/modules/school/uc/legacy-school.uc.ts
